fix: revert repos listing changes
