+++ .devcontainer/devcontainer.json
@@ -1,5 +1,5 @@
 {
-    "name": "AgentOps Dev Container",
+    "name": "Agentic Index Dev Container",
     "build": {
         "dockerfile": "Dockerfile"
     },
