+++ agentic_index_cli/__init__.py
@@ -1,1 +1,1 @@
-from agentops_cli import *  # re-export for backward compatibility
+# Agentic Index CLI package
