+++ agentic_index_cli/plot_trends.py
@@ -1,1 +1,7 @@
-from agentops_cli.plot_trends import *
+"""Placeholder for future trend plotting logic."""
+
+def main() -> None:
+    print("TrendGrapher is not yet implemented")
+
+if __name__ == "__main__":
+    main()
