+++ agentic_index_cli/prune.py
