+++ agentic_index_cli/ranker.py
@@ -1,1 +1,4 @@
-from agentops_cli.ranker import *
+from scripts.rank import main
+
+if __name__ == "__main__":
+    main()
