+++ agentic_index_cli/scraper.py
@@ -1,1 +1,4 @@
-from agentops_cli.scraper import *
+from scripts.scrape import main
+
+if __name__ == "__main__":
+    main()
