--- agentops_cli/inject_readme.py
@@ -1,4 +1,0 @@
-from scripts.inject_readme import main
-
-if __name__ == "__main__":
-    main()
