--- agentops_cli/plot_trends.py
@@ -1,7 +1,0 @@
-"""Placeholder for future trend plotting logic."""
-
-def main() -> None:
-    print("TrendGrapher is not yet implemented")
-
-if __name__ == "__main__":
-    main()
