--- agentops_cli/ranker.py
@@ -1,4 +1,0 @@
-from scripts.rank import main
-
-if __name__ == "__main__":
-    main()
