--- agentops_cli/scraper.py
@@ -1,4 +1,0 @@
-from scripts.scrape import main
-
-if __name__ == "__main__":
-    main()
