+++ docs/METHODOLOGY.md
@@ -1,3 +1,3 @@
 # Methodology
 
-This document outlines how AgentOps ranks repositories. It explains the scoring formula and the data sources used.
+This document outlines how Agentic Index ranks repositories. It explains the scoring formula and the data sources used.
