+++ docs/index.md
@@ -1,3 +1,3 @@
-# AgentOps Documentation
+# Agentic Index Documentation
 
-Welcome to the AgentOps documentation site. Here you'll find an overview of the project, usage instructions, and API reference material.
+Welcome to the Agentic Index documentation site. Here you'll find an overview of the project, usage instructions, and API reference material.
