+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: AgentOps
+site_name: Agentic Index
 theme: material
 plugins:
   - search
