+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = agentops
+name = agentic-index
 version = 0.1.0
-description = CLI for AgentOps
+description = CLI for Agentic Index
 
 [options]
 packages = find:
@@ -11,4 +11,4 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    agentops = agentops_cli.__main__:main
+    agentic-index = agentic_index_cli.__main__:main
