+++ tests/test_helpers.py
@@ -4,7 +4,7 @@ from pathlib import Path
 # ensure project root is on the path
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from agentops_cli import helpers
+from agentic_index_cli import helpers
 
 
 def test_add():
