+++ web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "agentops-web",
+  "name": "agentic-index-web",
   "version": "1.0.0",
   "scripts": {
     "build": "rm -rf dist && mkdir -p dist/data && cp index.html dist/index.html && cp data/repos.json dist/data/repos.json"
